Allow use of `$_` for unused variables in PHPMD
Creates a new rule extending from the base CamelCaseVariableName rule,
which simply returns true if the variable being tested is `$_`. In all
other scenarios, we fall back to the default rule handling.

There is already an `allow-underscore` option, but that only applies for
an underscore *prefix*, **not** cases where `_` is the only character.