"Fix" unused parameter warnings from PHPMD
This suppresses errors for unused `$attribute` and `$request` method
parameters in all rules and the register controller. It would be great
if PHPMD understood the same `@unused-param` syntax as Phan, but that's
a project for some day, some time in the distant future.

Also adds an exception to the `UnusedLocalVariable` rule for `$_`.