Prevent tests breaking when dummy dir exists
These directories get moved or deleted when the test passes so, when
they fail and get left behind, we don't want the next run to break
purely as a result of the last one failing. By checking `is_dir` first,
we can continue silently and only create when it doesn't exist.