Split project creation into multiple routes
This takes the support for creating a project with the application(s)
and/or redirect(s) pre-attached and moves it into two separate requests.

In order to allow for routes like '/api/projects/42/apps', this also
updates the `endpoint()` test helper to support replacing an `{id}` in
the base endpoint uri set on the corresponding test class.