Add progress modal on project creation
Creates a new ProgressModal component that's loaded with a title and
list of steps that will be completed, and updates the icon from 'minus'
to a 'check' when the `progress/progress` action's dispatched on a step.

Currently only used for creating a project, but the component has been
made rather generic so it can potentially be used elsewhere in future.