Force install of spatie/macroable 1.x
Macroable v2 was being required because I'm running install on php8, but
it drops support for php 7 entirely thus breaking install on 7.4.

Although spatie/ray allows v1 *or* v2, composer doesn't seem to want to
try for v1 on 7.4 even when I explicitly add it to the composer.json so
instead of messing about this just locks it to 1.x, since php8 support
has been added since 1.0.1 anyway so it's probably fine.