Fix user issue introduced in composer update
In v8.32.0, laravel/framework#36504 was merged which fixes null being
treated as true by the validator, but also enforces strict checking for
boolean values given to the required_unless rule. That's not necessarily
a bad thing, it just means we can't be passing `1` when we want `true`.

When we use `1`, user tests fail because while we pass `'user_group' =>
true` in the request data, the validator checks explicitly for `1` which
it doesn't see, so it says "gid is required unless user group is in 1".
We don't want that.