Split app/redirect observers to individual events
The code that actually works the magic is largely kept the same, with
the exception that `$app` and `$redirect` now have to be accessed via
the `$event` var instead, since handlers get an array of the event's
public properties, and that is either the Application or Redirect.

The idea behind splitting the App and Redirect creation/save handling
code into individual events is to make it easier for progress tracking.