Send progress events from project creation flow
Adds a new ProjectProgress event called from various steps in the
process so that progress can be reported on the fly with pusher/etc.

So that both apps *and* redirects' events can be easily handled by the
same code, this also adds the $project class property to the respective
ProjectFooSaved events, thus becoming part of the received payload.

For now we simply log events to the console when they're received on the
frontend, as the data is subject to change once the UI is figured out.