Update steps in realtime via progress events
Adds a new ProgressStep class that holds the name, text and status which
is used on the frontend to add and update steps on the fly, allowing us
to mark them as either skipped or complete depending how the task goes.

This new class entirely replaces the previous `$event->text` property,
though something similar could be added back in future for more granular
output from tasks such as cloning the project files or reloading nginx.