Delete ProjectProgress events with missing models
This fixes the `QUEUE_DRIVER` option in our phpunit config which didn't
didn't get updated properly when the rest of the config was upgraded.

With that, a `ModelNotFoundException` got thrown on 4 tests which all
`@depend` on a previous test that creates the Project. Not sure why it
only affects ProgressEvent and not Project{App,Redirect}Saved events,
but essentially it seems to refresh the db before the broadcast runs.

This works around it by discarding any ProjectProgress event that's
missing the Project model. It'd be better if broadcastAs/On had their
own tests using `assertPushedOn()` etc from the [Queue Fake] docs, but
right now I'm too lazy to do anything about it.

This also sets the broadcast connection to 'null' in the phpunit config
to make sure it won't attempt to send anything over pusher, but this
doesn't prevent Laravel from calling broadcast{On,As}() for coverage.

[queue fake]: https://laravel.com/docs/8.x/mocking#queue-fake