Switch progress events to private channels
For this to work we need a custom authoriser for pusher so that we can
use axios to get it sending the right headers along. If I can be arsed
figuring out which headers those are, we might just be able to pass it
along to pusher directly.

Similarly this moves the Broadcast::routes() into routes/api.php, which
is suggested by the Sanctum docs. If we pass the /api prefix it could
potentially work back in the broadcast service provider, but now things
are working (ish) I'm kind of tempted to leave them that way.