Wrap dispatches in a couple `Promise.all()`s
By grouping all the project creation steps into two sets of tasks,
separated only by the extra consts needed for later ones, eslint is a
lot closer to being happy with the number of statements in `create()`.

Promise.all() returns an array with all the promise resolutions, so we
can couple it with array destructuring to have the first project
request grouped with all the others in the first half.

To save the rest of the lines, error handling moves to a new function.