Add phpinsights dependency
There's no recent release for this, all the php8 etc stuff is on the
master branch so it set dev-master automatically when I required it.

If we try to install on 7.4 though, composer locks psr/cache at 2.x by
default which requires php >= 8. To workaround that we lock it at 1.x.