Update GET /api/databases to return object list
If we want to add any other information, we need more than just a list
of strings. This updates listDatabases() to return a keyed array so that
the json we get from the API can be used as objects.

Since it's no longer a simple array, `in_array` isn't good enough to
check if a database exists so that's abstracted to `hasDatabase()`.