Use get/keyBy in place of firstWhere
We have to override the get method to be able to infer the right types,
but once we `keyBy('name')` (albeit via a callback) we're able to use
get/has/etc without the need for tedious wheres/filters everywhere.

The only downside is we have to call `toArray()` manually if we don't
want it keyed in the response. We have to use `array_values()` there
because calling `$this->values()->toArray()` would kill everything.