Refactor for (slightly) cleaner unit tests
This drops both the `dbal()` and `connection()` methods, moving the
connection logic into the constructor so that we can pass in the new
dummy Connection class containing a fake schema manager that makes its
changes on an internal array rather than hitting the real database.

Unit tests are updated to reflect new functionality (as well as finally
renaming the DatabaseTest file), but feature tests need some work still.