Use mocks in feature test, inject SchemaManager
This updates the DatabaseManager to take an optional schema manager
rather than the connection class, to avoid issues where leaving the
resolution to `$connection->getSchemaManager()` can result in null.

The dummy connection is now a FakeSchemaManager which is used in the
unit tests where we can pass dependencies directly, whereas Mockery is
now used throughout the database creation feature tests.