Add database viewer (skeleton) page
No actual rows yet, but this adds a new page ready for table data to be
plugged in, currently with a column for the table name. The database
list gets updated to link each row to its respective viewer page, and it
all gets wrapped in a `<tbody>` so that the `selectable` class can work.