Don't automatically stash and pop around fixes
I have a horrible habit of using git add --patch to partially stage
files, committing a big batch of changes in chunks that make sense.

Unfortunately, that causes havoc with the idea of stashing, fixing and
popping automatically before commit. In reality, it will often stash,
apply fixes, then throw its toys out the pram with a merge conflict.

A lot of the time, this results in committing changes that were unstaged
previously, as the conflict resolution screws up what was staged before.