Second attempt at auto-fix
Rather than stashing, fixing and popping files, the idea behind this
method is to fix *only* those files that are completely staged, ie with
no additional changes in the working tree prior to fixing.

We still have to stash so that we can identify whether the files to be
committed will need fixing, and not just some unstaged file we don't
care about. If we reset anything that's left after adding all the "safe"
fixes to the commit, the stash can be popped entirely free of conflicts.