Allow commit even when files get stashed
As long as there are no dirty files after safe fixes are added to the
commit, it should be safe to continue with the commit after stashing any
changes that were initially backed up prior to fixing the staged ones.

This also adds a bit more logging to the remaining steps purely so it's
a bit easier to see what's happening... in case anything goes wrong.