Fix changes being lost in non-PHP files
With a single PHP file added and a non-PHP file with unstaged changes,
if the PHP file has fixes applied then the process would continue as if
there were still fixes on other files because the diff wasn't filtered.

As a result, the unstaged changes were lost because non-PHP changes
aren't considered when deciding if we should stash things. That also
means as a result that any time there *are* fixes to "non-safe" PHP
files, *all* unstaged non-PHP changes would be lost.

This fixes both issues by limiting the postFixDirty command to PHP files
(so it doesn't see a non-PHP file as a failed fix) and applying the fix
reset in a loop for each file directly, instead of a catch-all `.`.

Touch wood...