Fix install script aborting due to error
When reading from `/dev/urandom`, all is fine until it's piped through
`head` which cuts the stream short and throws a 141 signifying PIPEFAIL.

This presumably broke when `$(tr ... < /dev/urandom | head -c20)` was
changed to `$(</dev/urandom tr ... | head -c20)`, but I have no idea why
(or indeed if) it worked prior to that, or even why it was changed.
Could've been a false positive from shellcheck that's since been fixed?

tl;dr: If we add a test on the exit code being 141, we can bypass the
SIGPIPE error because it's 100% expected behaviour in this situation.