Fix login being required in new tabs
Since the `loggedIn` getter relies on there being a `user` object saved
in SessionStorage, opening links in a new tab results in a redirect back
to /login, even though the session may still be active and usable.

This resolves the issue by first attempting to re-fetch the profile
(which will set the `user` object back in SessionStorage) but *only*
when we're trying to fetch an authed page while seemingly "not authed".