Update framework files, drop proxy+cache packages
Removes the manual dependency on psr/cache from composer.json because
the lock was only needed for 1.x, prior to requiring php 8 or higher.
Also removes fideloper/proxy as laravel now includes it in core.

This also updates all nested package deps, as well as bumping
laravel/tinker to v2.6.2 which was released recently.