Adjust project save/toggle progress values
Prior to #435, enabling was part of deploying in the case of apps, or
done with the nginx config step for redirects. As a result there was no
explecit progress percentage except when disabling an app.

That app-disabling percentage was then used for enabling *or* disabling
any project, be it an app or a redirect. The problem is apps are at 85%
once cloned, and enabling happens after. A redirect would be 70%.

This commit allows for different progress when en/disabling, and updates
the progress when applying redirect configs to reflect the extra step.