Don't redirect when project errors are fixable
This adds a second `sui-button` that will hide the progress modal rather
than simply linking to the projects list if response errors can be fixed.

When there are validation errors, instead of a "Continue" button there's
now a red "Fix errors" button. Other errors will instead display a grey
button with the text "Back to Projects" so the intent is clear.