Fix duplicate progress step keys on second init
This replaces `setTitle` with `setup` so that, when called by the
progress/load action, the button and steps are reset to defaults.

Fixes an issue where retrying project creation after fixing validation
errors would simply add a second set of steps, causing Vue to break
because of duplicate keys for `project.create` and `app|redirect.save`.