Disable project name when app/redirect has errors
If you submit the project form with an invalid domain or repo, but the
project name is fine, the second submission will fail because the
project was already created the first time.

This keeps track of the created project ID and—if it's set—the create
action will simply set it completed and skip the actual creation,
instead moving straight on to the app/redirect after monitor setup.