Fix source selector not using old form values
The NewProject component wasn't using `v-model` before, so while the
`selected` handler could set values for the POST request, if there are
any validation errors then the old values weren't able to be reused.

This updates the SourceSelector component to have a `value` prop which
we access through a computed property to avoid prop mutation errors.

Also fixes an issue with the branch dropdown not showing the default
placeholder text, which requires a default value of `null` to be set.