Handle collection transforming
While the DTO can convert an array of DTOs to an array, it isn't able to
detect when we're using a collection implementing the toArray() method.

This commit overrides `parseArray` to solve that issue, and also sets a
constructor to avoid having the base DTO trying to set $tables to null.