Implement Arrayable to fix tests
The get method of the Database collection isn't able to find results
because the parent `toArray` defined in the `EnumeratesValues` trait
will only call `toArray` recursively if it implements `Arrayable`.

Also adds an extra assertion to the toArray and WithTables tests to make
sure that table arrays are the same regardless where `toArray` is called.