Dispatch new ProjectSaved event to fix toggling
Enabling or disabling an "app" actually happens on the project, but
saving the project doesn't trigger the usual events that would happen
when you're saving a project's app or redirect instead.

This adds a new `ProjectSaved` event similar to `ProjectAppSaved`, used
to trigger the toggle and reload bits when a project is saved by itself.