Fix tests and psalm/phan complaints
Psalm can't detect the type of `$project->applications` despite it being
in the docblock on the `Project` model, and laravel collections don't
exactly help matters here either, returning `mixed` from `first()`.

This fixes that issue with a bunch of ugly checks in the `getApp` and
`getRedirect` methods. For some reason projects aren't always returning
with an `applications` or `redirects` property now either, so we fix
that by passing `->with([...])` in the chain for affected tests.