Test for redirects and ensure config is removed
Tests can be uncertain if the nginx config and symlink aren't removed
after a run, so this does exactly that and asserts to be extra sure.

Also updates the test to use a Data Provider which gives us an easy way
to run it twice (once for an app, again for a redirect) without having
to duplicate the entire test method just to change data and endpoint.