Simplify the make-it-happy code
After all that nonsense, coverage was complaining that the array lines
weren't being hit in tests, which probably means despite the hint that
arrays aren't that likely to happen.

With that in mind, this simplifies the hinting hacks drastically.
Asserting alone results in mixed Collection items, but hinting only
`Collection<Application>` makes Phan complain that it's redundant so
instead we have to do both *and* include `array<Application>` too.