Rework missing data checks
Adds a new `MissingProjectData` exception to clean up the calling code
somewhat when it comes to sending the 'skipped' progress events.

Templates (and the `Redirect` model) now set a `requiredNginxData`
property to a list of attributes and their "friendly" names. This is
then used by the `RequiresNginxData` trait to throw if necessary.