Add dropdown to choose between custom or archive
Moves the redirect type dropdown to the top of the form alongside a new
'Rule Template' dropdown with options for either custom or WBM archives.

With this, the fields related to WayBack Machine archive date/time now
use the full form width and will not be visible if 'Custom' is chosen.