Appending `$request_uri` should be optional
Rather than forcing it in the nginx template, this commit updates the
project form to include an option that will append it as part of the
target when the redirect is initially created.

Probably still not the perfect solution, but until everything is done as
'server' blocks/options rather than "apps and redirects", this'll do.