Fix Composer installation failing in cloud-init
When the install script is run by cloud-init, it doesn't seem to set the
`$HOME` environment variable, causing the Composer install script to
fail with a 'missing HOME or COMPOSER_HOME' error.

By temporarily setting `COMPOSER_HOME=/tmp/composer` we directly set the
path which is used to store its keys. Once servidor is cloned, the
composer dir is moved to where it would be with `HOME=/var/servidor`.

We can't just use the latter because then Git wouldn't be able to clone
as the directory already exists. There are ways around that, but it's
really not worth the effort. Simply moving the composer keys is easier.