Fix permissions on some newly-made project dirs
When a project requires a user, causing it to get a docroot in that
user's homedir, the project is then created within *that* dir.

Problem is nginx can't see that project dir because the homedir doesn't
have o+x by default, so that's added in this commit. The pull method is
also updated to set the owner to the relevant user rather than www-data.

When creating the initial project dir, we now set `g+s` as well so that
the group will automatically be set to www-data when the cloning git.