Ensure config is defined
This fixes two bugs caused as the config field is `sometimes|required`.
When creating a project without SSL, you'd get a validation error
because although the config object is present, it's sent empty.

To ensure config is always set, this now sets `ssl` to `false` at mount
time, assuming it doesn't already have a value set in the value prop.

Also fixes a bug where config was being written to apps for redirects
because the check was done against the template **object** not name.