Fix creation values not being reset on success
For some reason the `DEFAULTS` const seems to get its value updated
somehow so even though it *was* used to reset things, its value was the
same as the project that had just been created (or was being discarded).

No idea how it was changing, even cloning with `{ ...DEFAULTS.project }`
didn't help, and nothing else seems to directly reference it. Still,
manually setting the values to something hard-coded is an easy fix.