Fix steps not being added on project creation
When saving a new project with an application of some sort (not a
redirect), many of the events won't have their step added despite the
event itself still being sent through Pusher.

This updates the affected "jobs" to trigger the start progress event,
which gives the progress modal a bit of a kick up the rear.