Fix progress being set too high/low
The app/redirect creation step, despite being among the first to fire,
doesn't complete until it's been enabled at the end of the process. This
causes the progress to skip back to 25/40% so this adjusts them to 95%,
leaving only the nginx reload to set 100% once that finishes after.

This commit also ensures progress is only set when a step is completed,
fixing an issue where progress would jump as soon as a task started. To
avoid the early nginx reload from prematurely setting 100% progress, the
stepCompleted action is also update to require a maximum increase,
meaning only the second config reload will actually set it to 100%.