Fixes various dark-mode bugs in the project wizard
Applies the inverted attribute to all project creation forms and adds
some extra styles for the steps, dropdowns and checkbox toggle switches.

For buttons, positive is replaced with `color="green"` in dark mode
because the inverted positive is an ugly white-bordered green button and
I'm too lazy to just make it look more like the positive primary button.

Also fixes the placeholder segment on the main project list page in
addition to the actual projects list and the search input at the top.
The progress modal also lacked inverted styles so those are added too.