Fix lack of custom repo formatter
The custom repository provider only had a name and text set which meant
the text component of the confirmation screen would get an undefined
error when trying to replace the '%REPO%' part with the real string.

This sets the custom repo formatter to just '%REPO%' to avoid issues,
and also updates the url input to use the same value as a normal github
or bitbucket repository would use which avoids needing extra handling.